requirements.txt
# ============================================
# AI 内容脱水日报 - 依赖清单
# ============================================

# === 爬虫层 ===
curl_cffi>=0.7.0              # 轻量级抓取，支持 impersonate
playwright>=1.40.0            # 重量级抓取，JS渲染
playwright-stealth>=1.0.6     # 绕过检测

# === 解析层 ===
beautifulsoup4>=4.12.0        # HTML解析
lxml>=5.0.0                   # 高性能解析器
readability-lxml>=0.8.1       # 正文提取算法

# === AI 层 (多模型支持) ===
anthropic>=0.39.0             # Claude API
openai>=1.50.0                # OpenAI & Azure OpenAI
httpx>=0.27.0                 # 自定义模型HTTP调用

# === 异步/调度 ===
aiohttp>=3.9.0                # 异步HTTP
tenacity>=8.2.0               # 重试机制
aiofiles>=23.2.0              # 异步文件操作

# === 配置/数据模型 ===
pydantic>=2.5.0               # 数据验证
pydantic-settings>=2.1.0      # 配置管理
pyyaml>=6.0.1                 # YAML解析

# === 输出层 ===
rich>=13.7.0                  # 终端美化
jinja2>=3.1.0                 # 模板引擎

# === Ray 分布式 (可选，取消注释启用) ===
# ray[default]>=2.9.0

# === 开发工具 ===
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
